+++ example/lib/app/modules/home/views/home_view.dart
@@ -114,37 +114,28 @@ class HomeViewState extends State<HomeView> {
                 ),
         ],
       ),
-      body: SizedBox(
-        width: 400,
-        child: ListView.separated(
-          padding: const EdgeInsets.all(10),
-          reverse: true,
-          separatorBuilder: (context, index) => const Divider(),
-          itemBuilder: (context, i) {
-            return VVoiceMessageView(
-              controller: controller.getVoiceController(voicesList[i]),
-              key: ValueKey(voicesList[i].id),
-              colorConfig: const VoiceColorConfig(
-                activeSliderColor: Colors.red,
-                notActiveSliderColor: Colors.grey,
-              ),
-              visualizerConfig: const VoiceVisualizerConfig(
-                useRandomHeights: true,
-                enableBarAnimations: false,
-              ),
-              textConfig: VoiceTextConfig(
-                  counterTextStyle: const TextStyle(
-                fontSize: 8,
-                fontWeight: FontWeight.w400,
-                height: .1,
-              )),
-              containerConfig: const VoiceContainerConfig(
-                borderRadius: 10,
-              ),
-            );
-          },
-          itemCount: voicesList.length,
-        ),
+      body: ListView.separated(
+        padding: const EdgeInsets.all(10),
+        reverse: true,
+        separatorBuilder: (context, index) => const Divider(),
+        itemBuilder: (context, i) {
+          return VVoiceMessageView(
+            controller: controller.getVoiceController(voicesList[i]),
+            key: ValueKey(voicesList[i].id),
+            colorConfig: const VoiceColorConfig(
+              activeSliderColor: Colors.red,
+              notActiveSliderColor: Colors.grey,
+            ),
+            visualizerConfig: const VoiceVisualizerConfig(
+              useRandomHeights: true,
+              enableBarAnimations: false,
+            ),
+            containerConfig: const VoiceContainerConfig(
+              borderRadius: 10,
+            ),
+          );
+        },
+        itemCount: voicesList.length,
       ),
     );
   }